add_property

+++ buffet2gether_home/lib/models/infoInGroup_model.dart
@@ -8,7 +8,7 @@ class InfoInGroup
   String imageUrl;
   int ageStart;
   int ageEnd;
-  int people;
+  double people;
   String dueTime;
   String gender;
