+++ buffet2gether_home/lib/models/infoInTable_model.dart
@@ -8,7 +8,7 @@ class InfoInTable
   String imageUrl;
   int ageStart;
   int ageEnd;
-  int people;
+  double people;
   String dueTime;
   String gender;
